# src/backend/routes/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Drivers Logbook Web Application API",
    "version": "1.0.0",
    "description": "API documentation for the Drivers Logbook Web Application."
  },
  "servers": [
    {
      "url": "http://localhost/",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/vehicles": {
      "get": {
        "summary": "Serves the vehicles page.",
        "description": "Returns the vehicles HTML page. Authentication is required.",
        "tags": [
          "Static Pages"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HTML file for the vehicles page."
          }
        }
      }
    },
...
    "/api/vehicles": {
      "get": {
        "summary": "Retrieves all vehicles.",
        "description": "Returns a list of vehicles. For admin users, all vehicles are returned; for managers and dispatchers, a filtered list is provided.",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of vehicles.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access."
          },
          "404": {
            "description": "No vehicles found or no vehicle assigned."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "summary": "Deletes all vehicles.",
        "description": "Removes all vehicles from the database. Requires admin API key.",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All vehicles deleted."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/vehicle/{id}": {
      "get": {
        "summary": "Retrieves a vehicle by ID.",
        "description": "Returns the vehicle data corresponding to the given ID.",
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The vehicle ID."
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "patch": {
        "summary": "Updates a vehicle.",
        "description": "Updates vehicle data based on the provided ID.",
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The vehicle ID."
          }
        ],
        "requestBody": {
          "description": "The vehicle data to update.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "brand": {
                    "type": "string"
                  },
                  "customName": {
                    "type": "string"
                  },
                  "manufacturer": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "year": {
                    "type": "number"
                  },
                  "licensePlate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Validation error."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "summary": "Deletes a vehicle.",
        "description": "Removes a vehicle by ID. Requires admin API key.",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The vehicle ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle deleted successfully."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/vehicle": {
      "post": {
        "summary": "Creates a new vehicle.",
        "description": "Adds a new vehicle to the database.",
        "tags": [
          "Vehicles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vin": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "year": {
                    "type": "number"
                  },
                  "licensePlate": {
                    "type": "string"
                  }
                },
                "required": [
                  "vin"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Vehicle already exists or validation error."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "Vehicle": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "licensePlate": {
            "type": "string"
          }
        }
      }
    }
  },
...
}